% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_python.R
\name{clean_python_settings}
\alias{clean_python_settings}
\title{Clean the Python Virtual Environment}
\usage{
clean_python_settings(envname = "r-RLoptimal")
}
\arguments{
\item{envname}{Python virtual environment name.}
}
\description{
Clean the Python Virtual Environment
}
