% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learn_allocation_rule.R
\name{learn_allocation_rule}
\alias{learn_allocation_rule}
\title{Build an optimal adaptive allocation rule using reinforcement learning}
\usage{
learn_allocation_rule(
  models,
  N_total,
  N_ini,
  N_block,
  Delta,
  sd_normal,
  optimization_metric = c("MAE", "TD", "power", "MS"),
  rl_models = models,
  rl_models_prior = NULL,
  rl_seed = NULL,
  rl_config = rl_config(),
  alpha = 0.025,
  selModel = c("AIC", "maxT", "aveAIC"),
  Delta_range = c(0.9, 1.1) * Delta,
  output_dir = format(Sys.time(), "\%Y\%m\%d_\%H\%M\%S"),
  output_base_dir = "allocation_rules",
  checkpoint_dir = "checkpoints",
  save_start_iter = 300L,
  save_every_iter = 100L
)
}
\arguments{
\item{models}{An object of class \link[DoseFinding]{Mods} specifying assumed
dose-response models.}

\item{N_total}{A positive integer value. The total number of subjects.}

\item{N_ini}{A positive integer value or vector. The number of subjects
allocated equally among dose groups initially.}

\item{N_block}{A positive integer value. The number of subjects allocated
adaptively in each round. This may differ from the actual clinical trials.}

\item{Delta}{A positive numeric value. The clinically relevant target effect.}

\item{sd_normal}{A positive numeric value. The standard deviation of the
observation noise.}

\item{optimization_metric}{A character value specifying the metric to
optimize. Possible values are "MAE" (default), "TD", "power", or "MS".
See Section 2.2 of the original paper for details.}

\item{rl_models}{An object of class \link[DoseFinding]{Mods}.}

\item{rl_models_prior}{A positive numeric vector.}

\item{rl_seed}{An integer value.}

\item{rl_config}{A list.}

\item{alpha}{A positive numeric value. The significance level. Default is 0.025.}

\item{selModel}{A character value specifying the model selection criterion
for dose estimation. Possible values are "AIC" (default), "maxT", or
"aveAIC". See \link[DoseFinding]{MCPMod} for details.}

\item{Delta_range}{A numeric vector of length 2.}

\item{save_every_iter}{}
}
\value{
An \link{AllocationRule} object.
}
\description{
Build an optimal adaptive allocation rule using reinforcement learning
}
