% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_python.R
\name{setup_python}
\alias{setup_python}
\title{Setting up a Python Virtual Environment}
\usage{
setup_python(envname = "r-RLoptimal")
}
\arguments{
\item{envname}{Python virtual environment name.}
}
\description{
Setting up a Python virtual environment for the Ray package, which includes
the RLlib library for reinforcement learning.
}
